chore: use OTLP_VERSION file to specify protocol version

+++ scripts/generate-protocol.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+set -e
+
 if ! type -P protoc >/dev/null 2>&1; then
   printf 'protoc is not available' 1>&2
   exit 1
@@ -14,11 +16,19 @@ PROTO_LENS=$(type -P proto-lens-protoc)
 
 SOURCE_PATH="$(git rev-parse --show-toplevel)"
 
-cd "${SOURCE_PATH}/protocol/" || exit 1
+cd "${SOURCE_PATH}/protocol/"
+
+OTLP_VERSION=$(cat OTLP_VERSION 2>/dev/null)
+
+if ! test -n "$OTLP_VERSION"; then
+  printf 'OTLP_VERSION file not available or version not specified\n'
+  exit 1
+fi
 
 # clone opentelemetry-proto repository
 test -e opentelemetry-proto && rm -fr opentelemetry-proto
 git clone -q https://github.com/open-telemetry/opentelemetry-proto.git
+git -C opentelemetry-proto reset --hard "$OTLP_VERSION" 1>/dev/null
 
 # generate Haskell opentelemetry protocol files
 find ./opentelemetry-proto/ -type f -name \*.proto -print0 |
